Debug Game#play
The cells kept track of in Player#slots_played were direct references to the cells themselves that got modified each time the player played into the cell. Player#add_slot was modified to fix this.

The game now plays as intended and stops when at least one of the following conditions are met; player 1 wins, player 2 wins, or the board is full.